Final_agent/requirements.txt
streamlit>=1.34
langgraph>=0.2.26
openai>=1.30
tavily-python>=0.3.3
pulp>=2.7
python-dotenv>=1.0
